pyproject.toml
[project]
name = "open-aasx-index"
version = "0.1.0"
description = "Automated AASX discovery, verification, and publishing system"
readme = "README.md"
license = "CC0-1.0"
requires-python = ">=3.11"
authors = [
    { name = "Open AASX Index Contributors" }
]
keywords = ["aasx", "asset-administration-shell", "aas", "industry-4.0", "digital-twin"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: CC0 1.0 Universal (CC0 1.0) Public Domain Dedication",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]

dependencies = [
    "aas-test-engines>=1.0.3",
    "basyx-python-sdk>=2.0.0",
    "requests>=2.31",
    "pyyaml>=6.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
    "responses>=0.25",
    "types-requests>=2.31",
    "types-PyYAML>=6.0",
]

[project.scripts]
harvest = "harvest.__main__:main"

[project.urls]
Homepage = "https://github.com/open-aasx-index/open-aasx-index"
Repository = "https://github.com/open-aasx-index/open-aasx-index"
Issues = "https://github.com/open-aasx-index/open-aasx-index/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["harvest"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["harvest"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = [
    "aas_test_engines.*",
    "basyx.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
